Refactor visits by reference test
